requirements.txt
h5py
wandb
numpy
pandas
pillow
gdown
plotly
einops
matplotlib
scikit-learn
scikit-survival
tqdm
wholeslidedata
seaborn==0.12.0
opencv-python==4.4.0.46
torchvision==0.13.0
